ajuste de subtotal al presupuesto y fixed al momento de registrar el costo del material

+++ database/migrations/2025_01_18_164924_create_materiales_inmuebles_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string("numero_identificacion",20);
             
             $table->decimal("costo_material");
-            $table->decimal("subtotal");
+            $table->unsignedBigInteger("subtotal");
             $table->decimal("cantidad_material");
             $table->string("codigo_proyecto",10);
