+++ database/migrations/2025_01_18_170052_create_asignaciones_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->unsignedInteger("consecutivo");
             
             $table->decimal("costo_material");
-            $table->decimal("subtotal");
+            $table->unsignedBigInteger("subtotal");
             $table->decimal("cantidad_material");
             $table->string("codigo_proyecto",10);
